Reduce calls to Patch advisories to a single request (#344)
* Reduce calls to Patch advisories to a single request

The subtotals for the three advisory types - bugs, security and
enhancements - are available in the `meta` structure of the request.
This means that we do not have to make three requests to the Patch API
in order to get the data necessary to build the Advisory Types pie
chart.

This work basically generates a single request and stores the subtotals
data.  The subtotal for each advisory type are then pulled out in the
PatchManagerCard component.

Signed-off-by: Paul Wayper <paulway@redhat.com>

* Update src/SmartComponents/PatchManager/PatchManagerCard.js

Handle initial load when advisories is undefined using an object destruction.

Co-authored-by: Muslimjon <59481011+mkholjuraev@users.noreply.github.com>

* Removing extra assignments

Signed-off-by: Paul Wayper <paulway@redhat.com>

---------

Signed-off-by: Paul Wayper <paulway@redhat.com>
Co-authored-by: Muslimjon <59481011+mkholjuraev@users.noreply.github.com>